Reworked the widget updating pipeline
After all models have been updated, the model thread sends a signal to the main window to update all widgets that depend on the changes in the model.
This makes the model thread only block once all relevant changes have been determined and the only outstanding changes are required to be done by the GUI Thread.
With that, all models should queue necessary changes required to be done in the GUI thread instead of emitting blocking update signals when called by the model threads update function.

The NetworkGraphWidget is currently the only user of this pipeline.