Rewrote update functionality of models
Removed the unnecessary passing of the current cycle in the update() function, as the BaseModel holds the current cycle.
Added the updateWidget() function, which is called on the GUI thread after the model thread has completed updating the models. The function should be used to complete tasks that have to be done in the GUI that were detected in the model update.
Some widgets were moved to this system already, while the Qt-Model based Views still need to be converted.