Improved plugin selector
The plugin selector will now save the state of all activated plugins when closing the window.
When opening the window, the plugins will be activated depending on the saved settings.
Updating the plugin working set (Opening the editor, updating the profile) will also trigger this. This can be changed by making this a startup-only procedure.